Fix nav bar spec
The current spec reports a false-positive due to the expectations
running before the call to api/authenticate finished. The DOM doesn't
exist yet by the time the assertion is triggered so the spec passes.

This change modifies the spec so it waits for the network request to
finish and asserts on the existence of the `#nav-mobile` ul element. It
also updates the component conditional to only render the element if
the user is logged in